Allow header logo to be set using an env var
The path to the PNG to be used as the header logo
can now be set using an environment variable.
If the variable is not set, the path defaults to the
previously used one for the ESS.

To alleviate problems with different aspect ratios
in the logos, the CSS for the PDF header now uses a more
flexible way of setting the image size. Setting the
minimum height to 54px ensures logos fit in the current
template and maintains the same appearance as before this
change for the ESS logo.

Part of UserOfficeProject/stfc-user-office-project#201